T8: Sign Up User Form Page Fixes #9
